Build statically linked binaries
We were incorrectly linking against the shared libraries from
e2fsprogs-libs.  This is bad because ldiskfsprogs updates to library
components would not appear in the installed binaries.  To work around
this issue we build statically.  A better fix would be to build a
ldiskfsprogs-libs pakcage to dynamically link against.  This will take
more work than the current fix, so we'll live with this for now so we
can move forward with testing.

Also, disable Obsolets and Provides statements in the spec file when
building with --with-ldiskfsprogs for those statements which are only
applicable to e2fsprogs.

Closes #1.